+++ src/spring/azext_spring/_params.py
@@ -24,7 +24,8 @@ from ._validators_enterprise import (only_support_enterprise, validate_builder_r
                                      validate_build_env, validate_target_module, validate_runtime_version)
 from ._app_validator import (fulfill_deployment_param, active_deployment_exist,
                              ensure_not_active_deployment, validate_deloy_path, validate_deloyment_create_path,
-                             validate_cpu, validate_memory, fulfill_deployment_param_or_warning, active_deployment_exist_or_warning)
+                             validate_cpu, validate_build_cpu, validate_memory, validate_build_memory,
+                             fulfill_deployment_param_or_warning, active_deployment_exist_or_warning)
 from ._app_managed_identity_validator import (validate_create_app_with_user_identity_or_warning,
                                               validate_create_app_with_system_identity_or_warning,
                                               validate_app_force_set_system_identity_or_warning,
@@ -51,7 +52,9 @@ source_path_type = CLIArgumentType(nargs='?', const='.',
                                    arg_group='Source Code deploy')
 # app cpu and memory
 cpu_type = CLIArgumentType(type=str, help='CPU resource quantity. Should be 500m or number of CPU cores.', validator=validate_cpu)
-memort_type = CLIArgumentType(type=str, help='Memory resource quantity. Should be 512Mi or #Gi, e.g., 1Gi, 3Gi.', validator=validate_memory)
+memory_type = CLIArgumentType(type=str, help='Memory resource quantity. Should be 512Mi or #Gi, e.g., 1Gi, 3Gi.', validator=validate_memory)
+build_cpu_type = CLIArgumentType(type=str, help='CPU resource quantity. Should be 500m or number of CPU cores.', validator=validate_build_cpu)
+build_memory_type = CLIArgumentType(type=str, help='Memory resource quantity. Should be 512Mi or #Gi, e.g., 1Gi, 3Gi.', validator=validate_build_memory)
 
 
 # pylint: disable=too-many-statements
@@ -216,7 +219,7 @@ def load_arguments(self, _):
                    validator=validate_create_app_with_user_identity_or_warning,
                    help="Space-separated user-assigned managed identity resource IDs to assgin to an app.")
         c.argument('cpu', arg_type=cpu_type, default="1")
-        c.argument('memory', arg_type=memort_type, default="1Gi")
+        c.argument('memory', arg_type=memory_type, default="1Gi")
         c.argument('instance_count', type=int,
                    default=1, help='Number of instance.', validator=validate_instance_count)
         c.argument('persistent_storage', type=str,
@@ -341,7 +344,7 @@ def load_arguments(self, _):
 
     with self.argument_context('spring app scale') as c:
         c.argument('cpu', arg_type=cpu_type)
-        c.argument('memory', arg_type=memort_type)
+        c.argument('memory', arg_type=memory_type)
         c.argument('instance_count', type=int, help='Number of instance.', validator=validate_instance_count)
 
     for scope in ['spring app deploy', 'spring app deployment create']:
@@ -377,6 +380,10 @@ def load_arguments(self, _):
                 'container_args', help='The arguments of the container image.', nargs='*', arg_group='Custom Container')
             c.argument(
                 'build_env', build_env_type)
+            c.argument(
+                'build_cpu', arg_type=build_cpu_type, default="1")
+            c.argument(
+                'build_memory', arg_type=build_memory_type, default="2Gi")
 
     with self.argument_context('spring app deploy') as c:
         c.argument('source_path', arg_type=source_path_type, validator=validate_deloy_path)
@@ -388,7 +395,7 @@ def load_arguments(self, _):
         c.argument('skip_clone_settings', help='Create staging deployment will automatically copy settings from production deployment.',
                    action='store_true')
         c.argument('cpu', arg_type=cpu_type)
-        c.argument('memory', arg_type=memort_type)
+        c.argument('memory', arg_type=memory_type)
         c.argument('instance_count', type=int, help='Number of instance.', validator=validate_instance_count)
 
     with self.argument_context('spring app deployment') as c: