+++ src/spring/HISTORY.md
@@ -4,6 +4,7 @@ Release History
 ---
 * New command `az spring create` has new argument "--ingress-read-timeout" to set ingress read timeout when create Azure Spring App.
 * New command `az spring update` has new argument "--ingress-read-timeout" to update ingress read timeout for Azure Spring App.
+* Command `az spring app deploy` and `az spring app deployment create` has new argument "--build-cpu" and "--build-memory" to set cpu and memory during build process.
 
 1.0.0
 ---